# Remove unnecessary `Arg` instantiation in `cuda/function.pyx`